rename sambanova package to maintain authorship
