--- llama-index-integrations/llms/llama-index-llms-sambanova/llama_index/llms/sambanova/__init__.py
@@ -1,3 +1,0 @@
-from llama_index.llms.sambanova.base import SambaNovaCloud
-
-__all__ = ["SambaNovaCloud"]
