+++ llama-index-integrations/llms/llama-index-llms-sambanovacloud/llama_index/llms/sambanovacloud/__init__.py
@@ -1,0 +1,3 @@
+from llama_index.llms.sambanovacloud.base import SambaNovaCloud
+
+__all__ = ["SambaNovaCloud"]
