+++ llama-index-integrations/llms/llama-index-llms-sambanovacloud/pyproject.toml
@@ -9,7 +9,7 @@ skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"
 
 [tool.llamahub]
 contains_example = false
-import_path = "llama_index.llms.sambanova"
+import_path = "llama_index.llms.sambanovacloud"
 
 [tool.llamahub.class_authors]
 SambaNovaCloud = "Pradeep"
@@ -22,7 +22,7 @@ python_version = "3.8"
 
 [tool.poetry]
 authors = ["Your Name <you@example.com>"]
-description = "llama-index llms sambanova integration"
+description = "llama-index llms sambanova cloud integration"
 name = "llama-index-llms-sambanovacloud"
 readme = "README.md"
 version = "0.3.1"
